:recycle: 서버에서 사용하는 타입들 분리
