+++ server/src/controller/passport/naver.ts
@@ -1,14 +1,7 @@
 import NaverStrategy from 'passport-naver';
 import User from '@models/User';
 import { NAVER_ID, NAVER_SECRET, OAUTH_CALLBACK_URL } from '@src/constant';
-
-interface profileType {
-  email: string;
-  nickname: string;
-  profile_image: string;
-  age: number | undefined;
-  birthday: any;
-}
+import type { NaverProfileType } from '@src/types';
 
 const naverStrategy = new NaverStrategy.Strategy(
   {
@@ -18,7 +11,7 @@ const naverStrategy = new NaverStrategy.Strategy(
   },
   async (accessToken, refreshToken, profile, cb) => {
     try {
-      const { email: naverId, nickname, profile_image: imgUrl } = profile._json as profileType;
+      const { email: naverId, nickname, profile_image: imgUrl } = profile._json as NaverProfileType;
 
       const existUser = await User.findOne({ naverId });
       if (existUser) return cb(null, existUser);
