+++ server/src/controller/socket/create.ts
@@ -1,27 +1,16 @@
-import { Socket } from 'socket.io';
-import type { roomType } from '@loader/socket';
-import type { TargetInfoType } from '@controller/socket/enter';
+import type { RoomType } from '@src/types';
 import { CREATE_REQUEST, CREATE_SUCCESS } from 'sooltreaming-domain/constant/socketEvent';
 import { STATUS_VOTE_NORMAL } from '@src/constant';
+import type { SocketPropType } from '@src/types';
 
-const createRoomCode = (rooms: roomType) => {
+const createRoomCode = (rooms: RoomType) => {
   while (true) {
     const code = Math.random().toString(16).substr(2, 5);
     if (!(code in rooms)) return code;
   }
 };
 
-const create = ({
-  io,
-  socket,
-  rooms,
-  targetInfo,
-}: {
-  io: any;
-  socket: Socket;
-  rooms: roomType;
-  targetInfo: TargetInfoType;
-}) => {
+const create = ({ io, socket, rooms, targetInfo }: SocketPropType): SocketPropType => {
   socket.on(CREATE_REQUEST, (user) => {
     const roomCode = createRoomCode(rooms);
     rooms[roomCode] = {
