+++ server/src/controller/socket/ticket.ts
@@ -1,23 +1,11 @@
-import { Socket } from 'socket.io';
-import type { roomType } from '@loader/socket';
-import type { TargetInfoType } from '@controller/socket/enter';
 import {
   TICKET_REQUEST,
   TICKET_SUCCESS,
   TICKET_FAILURE,
 } from 'sooltreaming-domain/constant/socketEvent';
+import type { SocketPropType } from '@src/types';
 
-const ticket = ({
-  io,
-  socket,
-  rooms,
-  targetInfo,
-}: {
-  io: any;
-  socket: Socket;
-  rooms: roomType;
-  targetInfo: TargetInfoType;
-}) => {
+const ticket = ({ io, socket, rooms, targetInfo }: SocketPropType): SocketPropType => {
   let roomCode = '';
   socket.on(TICKET_REQUEST, ({ code }) => {
     roomCode = code;
