+++ server/src/models/LiarGameKeywords.ts
@@ -1,8 +1,5 @@
 import { Schema, model } from 'mongoose';
-
-interface LiarGameKeywords {
-  keywords: string;
-}
+import type { LiarGameKeywords } from '@src/types';
 
 const LiarGameKeywordsSchema = new Schema<LiarGameKeywords>({
   keywords: String,
