+++ server/src/models/NicknameLog.ts
@@ -1,9 +1,5 @@
-import { Schema, model, SchemaDefinitionProperty } from 'mongoose';
-
-interface NicknameLog {
-  userId: SchemaDefinitionProperty<string>;
-  nickname: string;
-}
+import { Schema, model } from 'mongoose';
+import type { NicknameLog } from '@src/types';
 
 const nicknameLogSchema = new Schema<NicknameLog>(
   {
